chore: fix PRlinter dequeuing PRs because of CodeCov
If we don't know the result of the CodeCov results yet, we used to ask for changes, because
it prevents merging while the check might still fail in the future.

The following sequence of events happens because of this:

1. PR is ready to be merged (approved, everything passes)
2. Mergify enqueues it and merges from main
3. CodeCov needs to run again
4. PR linter requests changes because CodeCov result is uncertain
5. Mergify dequeues the PR because PR linter requests changes

This looks very confusing and noisy, and also will never fix itself, so the PR ends up unmerged.

The better solution would probably be not to do a "Request Changes" review, but leave a comment
and create a GitHub "status" on the PR to say 'success/pending/failure', and make it required.
(https://github.com/aws/aws-cdk/issues/33136)

For now, not doing anything with a 'waiting' status is a smaller delta, and the race condition posed by it is
unlikely to happen given that there are much slower jobs that the merge is blocked on anyway.

See also #33136.