chore: prlinter crashes if it runs alongside itself (#33129)
There is a race condition between multiple runs of the PR linter: it
finds a review that it wants to dismiss, but if that already has been
dismissed by another PR linter running in parallel the API call fails
and the linter does too.

Catch this specific case.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>