chore(cli-integ): optionally acquire environments from the cdk allocation service (#33069)
Closes https://github.com/aws/aws-cdk/issues/32437

### Reason for this change

In preparation for migrating the integration tests to use the new allocation service. Once the service is deployed, we can set its endpoint in our codebuild jobs environment and thats it.

### Description of changes

Introduce an environment variable `CDK_INTEG_ATMOSPHERE_ENABLED`. When it evaluates to true, the integration tests will perform a request to the allocation service (using its dedicated client, added as a new dependency).

### Description of how you validated changes

Ran against a service instance deployed onto a personal account.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*