chore: make integ tests run on GitHub Actions
Migrate some changes back from the new testing repo. These changes are
necessary to make the tests run on GitHub Actions.

If we keep them here, in the future we can do a `cp -R` on the test
directory. If not, we'll have to do manual sorting on every copy
over, which is annoying and easy to make mistakes in.