chore(cli-integ): make it possible to run on GitHub Actions (#33175)
Migrate some changes back from the new testing repo. These changes are necessary to make the tests run on GitHub Actions.

If we keep them here, in the future we can do a `cp -R` on the test directory. If not, we'll have to do manual sorting on every copy over, which is annoying and easy to make mistakes in.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*