fix(custom-resources): incorrect IAM prefix generated for CloudWatch actions (#33078)
### Issue # (if applicable)

Closes #32968.

### Reason for this change

The mapping to look up the IAM prefix from a given service has a incorrect entry for Cloudwatch. It says Cloudwatch uses `monitoring` as the prefix but it is actually `cloudwatch` instead.

I cannot find any service that uses `monitoring` as prefix so I think it is safe to assume that nothing relies on the `monitoring` value. Therefore, there is no feature flag used in this PR.

### Description of changes

Updated the IAM prefix mapping.

### Describe any new or updated permissions being added

Updated the mapping to use correct IAM prefix.

### Description of how you validated changes

Updated unit tests to use AwsCustomResource with a Cloudwatch call.

Added integ test to use AwsCustomResource with a Cloudwatch call to tag an alarm and verify the tag is indeed added successfully.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*