chore(custom-resource-handlers): lint (#33204)
The linter in the custom-resource-handlers folder was never turned on, resulting in typos, inconsistencies, and potentially bugs. This PR does not fix any potential bugs. I have tracked them in separate issues.

There's more work to be done with the generated files that we are not linting right now, but this is better than nothing

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*