chore(eslint): lint jsdocs (#33202)
Adds a dependency to `eslint-plugin-jsdoc` and utilizes the following linter rules (for now):

```
    'jsdoc/require-param-description': ['error'],
    'jsdoc/require-property-description': ['error'],
    'jsdoc/require-returns-description': ['error'],
```

These ensure that no one leaving `@param` or `@returns` without a description. In most cases, these are auto-generated and then the contributor forgets to add an actual description. They are useless without a description, so I am just removing the tags entirely. 

Our official stance is that these jsdoc tags are _optional_, but if you write them, they must have a description. 

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*