fix(batch): support cfn parameters for managed compute environment properties minvcpus, maxvcpus, and spotbidpercentage (#32954)
### Issue # (if applicable)

Closes #32905.

### Reason for this change

The Aws Batch Managed Compute Environment L2 currently does not allow for CfnParameter / Tokens for numeric properties that perform validation at synth time. This is a common pattern in other L2's in AWS CDK to skip the validation checks if the property can support being a cloud formation parameter

### Description of changes

I updated the validation methods in ManagedComputeEnvironment to skip if the property being validated is a token. Affected properties are spotBidPercentage, minVCpus, and maxVCpus. The logic keeps performing any of the checks if enough information in any of the properties are there to do the check at synth time.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Added new unit tests

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*