chore: add integ test that validates full proxy traversal (#33140)
(Re-roll of #33092)

Add a CLI integ test to validate that `cdk deploy` works in a fully network-isolated environment, with only a proxy to go through.

This validates that no parts of the CLI setup ignore the proxy configuration, which would otherwise be hard to test.

We achieve the network isolation by running the code inside a Docker container where we use `iptables` to drop all network traffic that doesn't go through the Docker host, where we run a proxy.

I temporarily bumped the `tsconfig` `target` to try out the `using` syntax (didn't work out with Jest), but that caused some compiler errors around class member initialization that I fixed as well.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*