refactor(toolkit): new message level `result` (#33234)
### Reason for this change

The CDK CLI has certain semantic rules around which messages are written to which output stream (`stdout` or `stderr`). These rules are mostly generalizable based on the level of a message. However until now there was one major exception: The main results of a CLI command should always be written to `stdout`, regardless of level. In practice these messages were always attributed to the `info` level.

With the recent refactorings towards a `CliIoHost`, we needed to include an additional property `forceStdout` on `IoMessages`. This always was a bad crutch and unintended to stay.

### Description of changes

Removal of the `forceStdout` message property, in favor of a new explicit message level: `result`.
In terms of priority, a `result` is lower than an `error`, but higher than a `warn`. This is intuitive: A user that wants to ignore all warnings, will still want to see results.

Use result in the CLU and the new Toolkit class.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Existing and extended tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*