chore(toolkit): default IoHost supports prompting (#33177)
### Reason for this change

The `ClioIoHost` didn't implement `requestResponse` properly, it just returned the default value.
However since this implementation is supposed to be exactly what the CLI does, we need to implement prompting.

### Description of changes

Implements user prompting. This is not currently used yet by the CLI itself, we will enable this in follow-up ticket.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Added test cases.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*