feat(ec2-alpha): add Transit Gateway L2 (#32956)
### Issue # (if applicable)

Closes https://github.com/aws/aws-cdk/issues/17528

### Description of changes

Create new Transit Gateway L2 with the following constructs:
 * `TransitGateway`
 * `TransitGatewayRouteTable`
 * `TransitGatewayRoute`
 * `TransitGatewayVpcAttachment`
 * `TransitGatewayRouteTableAssociation`
 * `TransitGatewayRouteTablePropagation`

#### Important Design Decision
As described in the README, the CDK disables the creation of the default
route table by EC2 and instead the CDK will create a "custom" default
route table in its place. This is primarily because there is no way to
obtain the route table ID of the EC2 created default route table without
a custom resource.

The CDK will disable the `defaultRouteTablePropagation` and
`defaultRouteTableAssociation` properties on the L1 (when both are
disabled, EC2 does not create the default route table), but the
properties are still exposed on the CDK TransitGateway interface to
allow it to be toggled for the CDK created default route table. The
automatic association/propagation is being mimicked by CDK
implementation and not relying on the actual EC2 feature.

### Describe any new or updated permissions being added

n/a


### Description of how you validated changes

Unit + Integration tests to verify that the deployed resources behave as
expected.

### Checklist
- [x] My code adheres to the [CONTRIBUTING
GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and
[DESIGN
GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>