feat(ecs): support container version consistency (#32225)
### Issue # (if applicable)

Closes #32202.

### Reason for this change

Support the new ECS::TaskDefinition ContainerDefinition VersionConsistency property. This is a simple enabled/disabled flag.

### Description of changes

Add a simple enabled/disabled enum prop to the construct.

More consequentially, default the prop to disabled (instead of unset) if the container image is a CDK asset, for the reasons described in the comments.

### Description of how you validated changes

Unit and integration tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*