chore(lambda-python-alpha): make pipenv version a parameter for bundling (#32594)
Make pipenv version a parameter in the Dockerfile for python lambda bundling. It was originally hardcoded to 2022.4.8, which excluded newer features/fixes.

### Issue # (if applicable)

None.

### Reason for this change

Pipenv version was originally hardcoded to 2022.4.8, which excluded newer features/fixes. And that could not be changed without providing a new Dockerfile from scratch.

### Description of changes

Just made the pipenv version a new ARG in the Dockerfile. So that it can be provided from the outside.

It has the same default value as it used to have, so that the current behavior won't change.

### Describe any new or updated permissions being added

None

### Description of how you validated changes

Tested this same Dockerfile in my current project, providing the PIPENV_VERSION variable via CDK:

```typescript
import * as python from '@aws-cdk/aws-lambda-python-alpha';

const lambdaProps: python.PythonFunctionProps = {
  runtime: lambda.Runtime.PYTHON_3_11,
  // ...
  bundling: {
    buildArgs: {
      'PIPENV_VERSION': '2024.0.1',
    },
  }
};
```

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*