fix(integ-tests): http flattenResponse (#30361)
### Issue # (if applicable)

Closes #30327

### Reason for this change

There was a difference in the behavior of SDK and HTTP integration attribute extraction with the `getAtt` and `getAttString` methods. `awsApiCall` properly implemented and returned JSONPath-ish values by using a `flattenResponse` property. This PR adds the same functionality to `httpApiCall`

### Description of changes

Added an implemented `flattenResponse` in the `HttpHandler` custom resource

### Description of how you validated changes

Updated integ and unit tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*