docs: correct code example in README of aws-cdk-lib/pipelines (#33354)
The code example in question mixed us-west-1 and us-east-1 in the same resource, which does not make sense and may be confusing.

Replace mentions of us-east-1 with us-west-1.

### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

I was reading through the CDK docs for the pipeline constructs and I saw this typo in the code example, so I wanted to get it fixed.

### Description of changes

I edited a code example in a README file to correct a documentation inconsistency.

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

N/A. No changes to functional components.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*