chore(cli): tests don't run on GitHub Actions
The `CliIoHost` tests using `promptly` don't run properly on GitHub
Actions. The event emitting on `process.stdin` we used to do doesn't
work on GitHub Actions. I don't know the reason, it might be a new Node
version or something.

Switching to a fake input stream (`streams.PassThrough`) seems to solve
it on both platforms.