chore(cli): tests involving promptly don't work on GitHub Actions (#33381)
The `CliIoHost` tests using `promptly` don't run properly on GitHub Actions. The event emitting on `process.stdin` we used to do doesn't work on GitHub Actions. I don't know the reason, it might be a new Node version or something.

Switching to a fake input stream (`streams.PassThrough`) seems to solve it on both platforms.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*