fix(cli): do not print credentials refreshed by plugins (#33398)
### Issue #33394

Closes #33394

### Reason for this change

In https://github.com/aws/aws-cdk/pull/32354 we started logging credentials provided by a plugin to `stderr`. In https://github.com/aws/aws-cdk/pull/32708 we changed the log statement in a way that the credentials would be logged to `stdout` in CI environments (but still to `stderr` in all other environments).

### Description of changes

Do not log credentials.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

n/a

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*