chore: retain existing status in the priority board on github action trigger (#33402)
### Issue # (if applicable)

Closes N/A

### Reason for this change

The existing PR's in the prioritization board with updated status such as In `Progress`, `Assigned` is automatically reset to the `Ready` status when the github action workflow is triggered for the same PR. 

### Description of changes

In this PR, made changes to check if a PR is already in the board first and if it is then we maintain the existing status and won't update the status. If the PR is new, then it will be added to board and also the status will be set to `Ready`.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

* Update unit tests
* Ran locally in test repo

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*