fix(opensearchservice): wrong iops limit checks (#33401)
### Issue # (if applicable)

Closes #29711.

### Reason for this change

The actual IOPS limit depends on a few factors:

![Screenshot 2025-02-11 at 1 15 41 PM](https://github.com/user-attachments/assets/de6ab224-03d6-4573-9c9a-a5b14584fb66)

Therefore, It is not trivial to maintain the limit checks in CDK. In addition, when the the service changes the limit, the CDK outdated limits will be a blocker for users to use the new limits. All in all, this check creates blockers more than benefit.


### Description of changes

Removed the limit check on IOPS.

### Description of how you validated changes

Unit Test that were expecting the error was failing after the removal of the check. Removed those test.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*